Move to Rust 2024 and Update Dependencies (#75)
* Move to rust 2024 and update deps

* Unify edition to workspace

* Improve misc code

* Revert version bump

* Add f64 tests and split off test from lib.rs into a test file

* Make dependency versions more consistent

* Improve `create_shader_module`s docs

* Revert "Improve `create_shader_module`s docs"

This reverts commit c9a5bfef39cf63b0038d6af51f5ae186da9a85b7.

* Re-embed test mod in lib.rs

* Format code

With `cargo fmt`

+++ wgsl_to_wgpu/src/consts.rs
@@ -2,7 +2,7 @@ use proc_macro2::{Span, TokenStream};
 use quote::quote;
 use syn::Ident;
 
-use crate::{wgsl::rust_type, MatrixVectorTypes, ModulePath, TypePath};
+use crate::{MatrixVectorTypes, ModulePath, TypePath, wgsl::rust_type};
 
 pub fn consts<F>(module: &naga::Module, demangle: F) -> Vec<(TypePath, TokenStream)>
 where
@@ -16,10 +16,9 @@ where
             let path = demangle(t.name.as_ref()?);
             let name = Ident::new(&path.name, Span::call_site());
 
-            // TODO: Add support for f64 once naga supports it.
             let type_and_value = match &module.global_expressions[t.init] {
                 naga::Expression::Literal(literal) => match literal {
-                    naga::Literal::F64(v) => Some(quote!(f32 = #v)),
+                    naga::Literal::F64(v) => Some(quote!(f64 = #v)),
                     naga::Literal::F32(v) => Some(quote!(f32 = #v)),
                     naga::Literal::U32(v) => Some(quote!(u32 = #v)),
                     naga::Literal::I32(v) => Some(quote!(i32 = #v)),
